+++ braldahim/library/Bral/Batchs/CreationBuissons.php
@@ -48,9 +48,10 @@ class Bral_Batchs_CreationBuissons extends Bral_Batchs_Batch {
 			$where .= $or." (x_buisson = ".$r["x_eau"]. " AND y_buisson = ".$r["y_eau"].") ";
 		}
 
-		$buissonTable = new Buisson();
-		$buissonTable->delete($where);
-
+		if ($where != "") {
+			$buissonTable = new Buisson();
+			$buissonTable->delete($where);
+		}
 		Bral_Util_Log::batchs()->trace("Bral_Batchs_CreationBuissons - suppressionBuissonSurEau - exit -");
 		return $retour;
 	}
@@ -150,7 +151,7 @@ class Bral_Batchs_CreationBuissons extends Bral_Batchs_Batch {
 		for($i = 1; $i <= $aCreer; $i++) {
 			$x = Bral_Util_De::get_de_specifique($zone["x_min_zone"], $zone["x_max_zone"]);
 			$y = Bral_Util_De::get_de_specifique($zone["y_min_zone"], $zone["y_max_zone"]);
-				
+
 			usleep(Bral_Util_De::get_de_specifique(1, 1000000));
 			$this->insertDb($buissonTable, $idTypeBuisson, $x, $y, 0, 1);
 		}
